+++ src/main/java/com/atorres/nttdata/transactionmicroservice/model/dao/TransactionDao.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -15,8 +16,8 @@ public class TransactionDao {
     private String from;
     private String to;
     private String category;
-    private Double balance;
-    private Double comission;
+    private BigDecimal balance;
+    private BigDecimal comission;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "America/Lima")
     private Date date;
     private String clientId;
